+++ src/RFNode.cpp
@@ -30,7 +30,7 @@ RFNode::RFNode(
 
 RFNode::RFNode(
   size_t splitFeature,
-  double splitValue,
+  float splitValue,
   std::unique_ptr< RFNode > leftChild,
   std::unique_ptr< RFNode > rightChild
 ) {
@@ -57,7 +57,7 @@ void RFNode::setLeafNode(
 
 void RFNode::setSplitNode(
   size_t splitFeature,
-  double splitValue,
+  float splitValue,
   std::unique_ptr< RFNode > leftChild,
   std::unique_ptr< RFNode > rightChild
 ) {
@@ -73,9 +73,9 @@ void RFNode::setSplitNode(
 
 
 void RFNode::predict(
-  std::vector<double> &outputPrediction,
+  std::vector<float> &outputPrediction,
   std::vector<size_t>* updateIndex,
-  std::vector< std::vector<double> >* xNew,
+  std::vector< std::vector<float> >* xNew,
   DataFrame* trainingData
 ) {
 
@@ -83,7 +83,7 @@ void RFNode::predict(
   if (is_leaf()) {
 
     // Calculate the mean of current node
-    double predictedMean = (*trainingData).partitionMean(getAveragingIndex());
+    float predictedMean = (*trainingData).partitionMean(getAveragingIndex());
 
     // Give all updateIndex the mean of the node as prediction values
     for (
