+++ src/rcpp_cppBuildInterface.cpp
@@ -25,15 +25,15 @@ SEXP rcpp_cppDataFrameInterface(
 ){
 
   try {
-    std::unique_ptr<std::vector< std::vector<double> > > featureDataRcpp (
-        new std::vector< std::vector<double> >(
-            Rcpp::as< std::vector< std::vector<double> > >(x)
+    std::unique_ptr<std::vector< std::vector<float> > > featureDataRcpp (
+        new std::vector< std::vector<float> >(
+            Rcpp::as< std::vector< std::vector<float> > >(x)
         )
     );
 
-    std::unique_ptr<std::vector<double>> outcomeDataRcpp (
-        new std::vector<double>(
-            Rcpp::as< std::vector<double> >(y)
+    std::unique_ptr<std::vector<float>> outcomeDataRcpp (
+        new std::vector<float>(
+            Rcpp::as< std::vector<float> >(y)
         )
     );
 
@@ -74,7 +74,7 @@ SEXP rcpp_cppBuildInterface(
   bool replace,
   int sampsize,
   int mtry,
-  double splitratio,
+  float splitratio,
   int nodesizeSpl,
   int nodesizeAvg,
   int seed,
@@ -122,15 +122,15 @@ SEXP rcpp_cppBuildInterface(
   } else {
 
     try {
-      std::unique_ptr<std::vector< std::vector<double> > > featureDataRcpp (
-          new std::vector< std::vector<double> >(
-              Rcpp::as< std::vector< std::vector<double> > >(x)
+      std::unique_ptr<std::vector< std::vector<float> > > featureDataRcpp (
+          new std::vector< std::vector<float> >(
+              Rcpp::as< std::vector< std::vector<float> > >(x)
           )
       );
 
-      std::unique_ptr<std::vector<double>> outcomeDataRcpp (
-          new std::vector<double>(
-              Rcpp::as< std::vector<double> >(y)
+      std::unique_ptr<std::vector<float>> outcomeDataRcpp (
+          new std::vector<float>(
+              Rcpp::as< std::vector<float> >(y)
           )
       );
 
@@ -191,15 +191,15 @@ Rcpp::NumericVector rcpp_cppPredictInterface(
 
     Rcpp::XPtr< honestRF > testFullForest(forest) ;
 
-    std::vector< std::vector<double> > featureData =
-      Rcpp::as< std::vector< std::vector<double> > >(x);
+    std::vector< std::vector<float> > featureData =
+      Rcpp::as< std::vector< std::vector<float> > >(x);
 
-    std::unique_ptr< std::vector<double> > testForestPrediction (
+    std::unique_ptr< std::vector<float> > testForestPrediction (
       (*testFullForest).predict(&featureData)
     );
 
-    std::vector<double>* testForestPrediction_ =
-      new std::vector<double>(*testForestPrediction.get());
+    std::vector<float>* testForestPrediction_ =
+      new std::vector<float>(*testForestPrediction.get());
 
     Rcpp::NumericVector output = Rcpp::wrap(*testForestPrediction_);
 
@@ -215,13 +215,13 @@ Rcpp::NumericVector rcpp_cppPredictInterface(
 
 
 // [[Rcpp::export]]
-double rcpp_OBBPredictInterface(
+float rcpp_OBBPredictInterface(
     SEXP forest
 ){
 
   try {
     Rcpp::XPtr< honestRF > testFullForest(forest) ;
-    double OOBError = (*testFullForest).getOOBError();
+    float OOBError = (*testFullForest).getOOBError();
     return OOBError;
   } catch(std::runtime_error const& err) {
     forward_exception_to_r(err);
@@ -233,13 +233,13 @@ double rcpp_OBBPredictInterface(
 
 
 // [[Rcpp::export]]
-double rcpp_getObservationSizeInterface(
+float rcpp_getObservationSizeInterface(
     SEXP df
 ){
 
   try {
     Rcpp::XPtr< DataFrame > trainingData(df) ;
-    double nrows = (double) (*trainingData).getNumRows();
+    float nrows = (float) (*trainingData).getNumRows();
     return nrows;
   } catch(std::runtime_error const& err) {
     forward_exception_to_r(err);
